// src/main/interfaces/requestIf.ts
export interface rejoinGameAgainReqIf{
    tableId: string;
    userId: string;
    currentRound : number;
 }

 export interface rejoinGameAgainHandlerReqIf{
    data : rejoinGameAgainReqIf
 }